simplify interface, split up single file into 2, update docs
