+++ docs/runbooks/AWS_DEPLOYMENT_GUIDE.md
@@ -98,11 +98,12 @@ Use `infra/main.tf` to provision the AWS infrastructure:
 From the **repository root**:
 
 1. Ensure `public/config.js` matches the deployed API (the scripts query `jspsych-scroll-api` in `us-east-2` and compare exact URLs).
-2. **Dry run** (creates `s3_upload/<timestamp>/`, writes `manifest.json`, no S3 writes):
+2. **Dry run** (stage locally, then print intended S3 keys without uploading):
 
    ```bash
    uv python install 3.12
    uv sync
+   PYTHONPATH=. uv run python scripts/upload_to_s3/stage_public_for_s3.py
    PYTHONPATH=. uv run python scripts/upload_to_s3/upload_public_to_s3.py --dry-run
    ```
 
@@ -117,6 +118,7 @@ From the **repository root**:
    Or run the Python steps yourself:
 
    ```bash
+   PYTHONPATH=. uv run python scripts/upload_to_s3/stage_public_for_s3.py
    PYTHONPATH=. uv run python scripts/upload_to_s3/upload_public_to_s3.py
    PYTHONPATH=. uv run python scripts/upload_to_s3/verify_s3_upload.py
    ```
